Add a function to set the editor status message
The function is editorSetStatusMessage(). It takes a various amount of arguments hence why we need to use va_list, va_start, va_end. We use vsnprintf to store the status message into E.statusmsg. We set the time to the current time using time(NULL). We initialize the default status message to be a help message in main.