Draw the status message in the status message bar
Created a method to print the status message to the status message bar. First, we clear the message bar using the escape sequence \x1b[K. Then we only print the status bar as much as the screen columns allow. We also only print the status message if the message is less than 5 seconds old.