Add insert character
First checked if the index at which the character is being inserted is in bounds, if not insert the char at the end of the character. We also reallocated space for the row->chars so that we can insert the new character and null terminator. We then use memmove to move the old string to the new location (allows overlapping instead of memcpy). Then, we increment the size and put the character at the row at the index at. We then call editorUpdateRow to render the new row and update render and rsize fields.