Add editorInsertChar() function
We add a new section called editor operations which deals with mapping keypresses to various text editing operations. We also create a function called editorInsertChar(). We first check if we're at the end of the number of rows in the file. If so, we create a new row. Then, we call editorInsertChar() to insert a character on the line we are on. Then we increment E.cx to move the cursor to the right after we type in a char.