Add editorRowsToString()
Created a function that copies all of the strings in each row of the editor to a single string for writing to a file. First, we find the total length by iterating through all of the rows and incrementing total len by each row's size. Then, we malloc space for the new string. We then iterate through the rows again and copy each row to the buffer with new line characters in between each row. We then return the buffer.