Add quit confirmation
We require the user to press ^q 3 times before quitting if they try to close when they have unmodified changes. We use a static variable to keep track of this, and decrement it each time they press ^q. Then, if the quit times finally reaches 0, the program quits. If they press any other key, the quit time is reset to 3.