Add editor delete char
We check if the editor is able to delete the char using this method, and if it is, we call editorRowDelChar. If we're past the end of the file, we return. If we're at the very beginning of a line, we also just return, as there is nothing to backspace.