Add a method to append a string to a row for deleting purposes
First we realloc enough space to hold the row's chars and also the string that's being appended. Then we copy over the string to the end of the row's chars. Then, we increment the size to size + len, and null terminate the row. We then update the row and increment dirty.