Add the ability to delete a row
First, we check if we're at the very beginning of the file. If we are, we return, if we aren't if we're at the beginning of a line, we set the x position of the cursor to the end of the previous line, and then append the lines together. Then, we delete the current row, and decrement the y position of the cursor.