Add editorInsertNewline()
If we're at the beginning of a row, we just insert a new line at that position using editorInsertRow. Else, we split the line into two parts. We call editorInsertRow to create a new row on the next line with all the characters after E.cx. We then truncate the current row by reducing the size to E.cx, and then updating the row. We then increment the y cursor to put it on the next line, and set E.cx = 0.