Add editorPrompt()
This method creates a buffer that the user can type into using the status bar. The method continuously refreshes and sets the status message to the buffer as the user presses keys. If enter is inputted, we set the status message to be empty and return the input. Else, if the character is printable, we can append it to the buffer and then null terminate it.