Add editorRowRxToCx
This method converts an Rx (render cursor position) to a Cx (char cursor position)
We first loop through the chars string, and calculate the current rx value as we go. We stop when the cur_rx hits the given rx value and return cx at that position. This will give us the Cx for a given Rx.