Add incremental search functionality
We split editorFind into editorFind and editorFindCallback, where editorFindCallback is the callback function for editorPrompt.

When we call editorFind, we call editorPrompt with editorFindCallback as the callback function.
After each keypress in editorPrompt, we call editorFindCallback which loops through the editor and matches the current query, thus implementing incremental search.