Add logic for last_match and direction in editorFindCallback
If we don't have a last match, we start searching at the beginning of the file in the positive direction.
Current is the index of the current row we are searching. If there was a last match, it starts on the line after (or before if we're searching backwards).
The conditionals cause current to go from the end of the file back to the beginning of the file or vice versa, to allow a search to wrap around the end of a file and continue from the top or bottom.
When we find a match, we set last_match to current, so that if the user presses the arrow keys, we'll start the next search from that point.